Refactors s3Upload function into its own module.
